V2
add Balance Table
Correction of unidirectionnel association
add Complex Methods

+++ MycroA/src/main/java/tn/saturn/spring/repositories/ContractRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import tn.saturn.spring.entities.Contract;
 
-public interface ContractRepository extends CrudRepository<Contract,Long >{
+public interface ContractRepository extends CrudRepository<Contract,Integer>{
 
 }
